+++ src/main/java/br/nassau/floresta/Tigre.java
@@ -2,7 +2,7 @@ package main.java.br.nassau.floresta;
 
 public class Tigre extends Animal {
 	// Construtor para Tigre
-	public Tigre(double startX, double startY) {
+	public Tigre(int startX, int startY) {
 		super("Tigre", "Predador", 'T', 5, startX, startY);
 	}
 }
